feat(zephyr): split M1 tracer into tx/rx examples, verify 2-node LoRa
Replace the single symmetric app/ with a role-split examples tree: a developer
now gets a clean transmit-only node and a receive-only node instead of one image
that both sends and listens. The symmetric loop could also phase-lock (both nodes
transmitting and listening in lockstep, never hearing each other); dedicated roles
remove that failure mode and read far more clearly.

Layout (ports/zephyr):
- common/xiao_esp32s3_procpu.overlay - SX1262 pins, single source of truth for
  every example (applied via DTC_OVERLAY_FILE).
- common/lora_phy.h - the interop PHY contract (915 MHz / SF8 / BW125 / CR4-7 /
  preamble 8 / private sync / 20 dBm) in one place, so tx and rx - and Arduino
  nodes (D5) - cannot drift.
- examples/tx - builds a RadioMeshPacket with the co-compiled core and transmits
  on a 3 s cadence.
- examples/rx - receives continuously, parses via RadioMeshPacket(buffer), and
  re-checks the payload CRC32.

Hardware-verified on two XIAO ESP32-S3 + Wio-SX1262 boards (Zephyr 4.4.0):
- rm_tx: TX fcounter=N len=38 ret=0  (over the air; TX-done IRQ)
- rm_rx: RX len=38 ... topic=0x10 dataLen=3 crc=OK
The co-compiled M0 core serializes, transmits, receives, and parses a packet over
a real Zephyr LoRa link - M1's 2-node tracer bullet.

Housekeeping:
- Drop the bring-up debug scaffolding (boot delay, printk FATAL loop, forced LoRa
  DBG) from the examples.
- Document the gpio_esp32 BIT64 patch re-apply step in the README.
- Note loramac-node as unused under CONFIG_LORA_MODULE_BACKEND_NATIVE; removal
  deferred to a verified west-update + rebuild step.

M0 gates still green: pio test -e native (4/4), pio run -e seeed_xiao_esp32s3.
