+++ ports/zephyr/examples/rx/prj.conf
@@ -14,10 +14,8 @@ CONFIG_CPP_EXCEPTIONS=y
 # --- Heap for std::vector/std::string allocations (tune on-target if needed) ---
 CONFIG_COMMON_LIBC_MALLOC_ARENA_SIZE=16384
 
-# --- Logging (deferred + delayed flush so boot-time driver logs survive USB-JTAG enumeration) ---
+# --- Logging ---
 CONFIG_LOG=y
-CONFIG_LOG_PROCESS_THREAD_STARTUP_DELAY_MS=3000
-CONFIG_LORA_LOG_LEVEL_DBG=y
 CONFIG_PRINTK=y
 
 # --- Stack sizing (LoRa + STL on the main thread) ---
