+++ ports/zephyr/examples/tx/prj.conf
@@ -1,0 +1,25 @@
+# --- LoRa PHY stack (native in-tree SX126x backend; loramac-node is deprecated) ---
+CONFIG_LORA=y
+CONFIG_LORA_MODULE_BACKEND_NATIVE=y
+CONFIG_SPI=y
+CONFIG_GPIO=y
+
+# --- C++17 with full STL (RadioMeshPacket uses std::vector/std::array/std::string) ---
+CONFIG_CPP=y
+CONFIG_STD_CPP17=y
+CONFIG_REQUIRES_FULL_LIBC=y
+CONFIG_GLIBCXX_LIBCPP=y
+CONFIG_CPP_EXCEPTIONS=y
+
+# --- Heap for std::vector/std::string allocations (tune on-target if needed) ---
+CONFIG_COMMON_LIBC_MALLOC_ARENA_SIZE=16384
+
+# --- Logging ---
+CONFIG_LOG=y
+CONFIG_PRINTK=y
+
+# --- Stack sizing (LoRa + STL on the main thread) ---
+CONFIG_MAIN_STACK_SIZE=8192
+
+# --- RadioMesh neutral core (zephyr/Kconfig) ---
+CONFIG_RADIOMESH=y
